chore: move transactions list into it's own file
